ADMIN LOGIN PORTAL
This is the main page, where an admin would log in. It creates a PDO object and then sets up a query with the login information specified, if someone tries to access another page without logging in here first, they will be redirect back here. To successfully login, USER_ID = 1 and password = ADMIN.